+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2024-02-04
+
+### Added
+
+- Removed `completions` endpoint(s)
+- Removed `edits` endpoint(s)
+- Removed `fine-tunes` endpoint(s)
+- Moved `deleteModel` to model endpoints
+- Updated examples. and tests
+- Allow override of `OpenAIUrlBuilder::$apiVersion`
+
 ## [2.1.1] - 2024-01-11
 
 ### Added